EX-351: SBP to MSM conversion (#143)
* add stub for sbp to msm conversion

* Extract MSM helper functions

* Refactor MSM PRN helper functions

* Convert SBP obs sequence into MSM messages

* disable legacy RTCM observations

* Fix unused variables errors when asserts are not compiled

* Extract utils unit tests to separate suite

* Unit test MSM code and PRN conversion utilies

* Configure RTCM type (legacy/MSM) in rtcm_out_state, default to MSM5

* Add settes and getters for RTCM out settings

* Ignore invalid and RAIMed measurements

* Clean up the public header

* Clean up and refactor

* add sbp to MSM sanity test

* Fix range rate computation

* Add unit test for sbp->msm->sbp conversion, clean up interfaces

* Robustify msm_add_to_header/cell_mask and add unit tests

* Address remaining todos

* Select RTCM out mode with a single enum

* Fix RTCM frame length from u8 to u16

* Proper rounding of rough rate and rough range rate

* On encoding error, send empty buffer instead of msg without payload

* Fix MSM roundtrip unit test to generate also GLO obs

* Update librtcm to v0.2.34

+++ c/src/sbp2rtcm.c
@@ -23,7 +23,7 @@
 #include <libsbp/sbp.h>
 
 /* Write the RTCM frame to STDOUT. */
-static void cb_sbp_to_rtcm(u8 *buffer, u8 n, void *context) {
+static void cb_sbp_to_rtcm(u8 *buffer, u16 n, void *context) {
   (void)(context);
 
   ssize_t numwritten = write(STDOUT_FILENO, buffer, n);
